+++ bolt/internal/internal.go
@@ -51,6 +51,7 @@ func MarshalSource(s chronograf.Source) ([]byte, error) {
 		Password: s.Password,
 		URL:      s.URL,
 		Default:  s.Default,
+		Telegraf: s.Telegraf,
 	})
 }
 
@@ -68,6 +69,7 @@ func UnmarshalSource(data []byte, s *chronograf.Source) error {
 	s.Password = pb.Password
 	s.URL = pb.URL
 	s.Default = pb.Default
+	s.Telegraf = pb.Telegraf
 	return nil
 }
 
@@ -201,5 +203,25 @@ func UnmarshalAlertRule(data []byte, r *ScopedAlert) error {
 	}
 	r.SrcID = int(pb.SrcID)
 	r.KapaID = int(pb.KapaID)
+	return nil
+}
+
+// MarshalUser encodes a user to binary protobuf format.
+func MarshalUser(u *chronograf.User) ([]byte, error) {
+	return proto.Marshal(&User{
+		ID:    uint64(u.ID),
+		Email: u.Email,
+	})
+}
+
+// UnmarshalUser decodes a user from binary protobuf data.
+func UnmarshalUser(data []byte, u *chronograf.User) error {
+	var pb User
+	if err := proto.Unmarshal(data, &pb); err != nil {
+		return err
+	}
+
+	u.ID = chronograf.UserID(pb.ID)
+	u.Email = pb.Email
 	return nil
 }