+++ server/mux.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	"github.com/bouk/httprouter"
-	"github.com/influxdata/chronograf" // When julienschmidt/httprouter v2 w/ context is out, switch "github.com/influxdata/chronograf
+	"github.com/influxdata/chronograf" // When julienschmidt/httprouter v2 w/ context is out, switch
 	"github.com/influxdata/chronograf/jwt"
 )
 
@@ -94,14 +94,13 @@ func NewMux(opts MuxOpts, service Service) http.Handler {
 	router.DELETE("/chronograf/v1/layouts/:id", service.RemoveLayout)
 
 	// Users
-	/*
-		router.GET("/chronograf/v1/users", Users)
-		router.POST("/chronograf/v1/users", NewUser)
+	router.GET("/chronograf/v1/me", service.Me)
+	router.POST("/chronograf/v1/users", service.NewUser)
 
-		router.GET("/chronograf/v1/users/:id", UsersID)
-		router.PATCH("/chronograf/v1/users/:id", UpdateUser)
-		router.DELETE("/chronograf/v1/users/:id", RemoveUser)
-	*/
+	router.GET("/chronograf/v1/users/:id", service.UserID)
+	router.PATCH("/chronograf/v1/users/:id", service.UpdateUser)
+	router.DELETE("/chronograf/v1/users/:id", service.RemoveUser)
+
 	// Explorations
 	router.GET("/chronograf/v1/users/:id/explorations", service.Explorations)
 	router.POST("/chronograf/v1/users/:id/explorations", service.NewExploration)
@@ -133,7 +132,7 @@ func AuthAPI(opts MuxOpts, router *httprouter.Router) http.Handler {
 		opts.Logger,
 	)
 
-	router.GET("/oauth", gh.Login())
+	router.GET("/oauth/github", gh.Login())
 	router.GET("/oauth/logout", gh.Logout())
 	router.GET("/oauth/github/callback", gh.Callback())
 
@@ -152,44 +151,45 @@ func encodeJSON(w http.ResponseWriter, status int, v interface{}, logger chronog
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(status)
 	if err := json.NewEncoder(w).Encode(v); err != nil {
-		unknownErrorWithMessage(w, err)
+		unknownErrorWithMessage(w, err, logger)
 	}
 }
 
 // Error writes an JSON message
-func Error(w http.ResponseWriter, code int, msg string) {
-	e := struct {
-		Code    int    `json:"code"`
-		Message string `json:"message"`
-	}{
+func Error(w http.ResponseWriter, code int, msg string, logger chronograf.Logger) {
+	e := ErrorMessage{
 		Code:    code,
 		Message: msg,
 	}
 	b, err := json.Marshal(e)
 	if err != nil {
-		//log.Print("go-oidc: failed to marshal %#v: %v", e, err)
 		code = http.StatusInternalServerError
 		b = []byte(`{"code": 500, "message":"server_error"}`)
 	}
+
+	logger.
+		WithField("component", "server").
+		WithField("http_status ", code).
+		Error("Error message ", msg)
 	w.Header().Set("Content-Type", JSONType)
 	w.WriteHeader(code)
 	w.Write(b)
 }
 
-func invalidData(w http.ResponseWriter, err error) {
-	Error(w, http.StatusUnprocessableEntity, fmt.Sprintf("%v", err))
+func invalidData(w http.ResponseWriter, err error, logger chronograf.Logger) {
+	Error(w, http.StatusUnprocessableEntity, fmt.Sprintf("%v", err), logger)
 }
 
-func invalidJSON(w http.ResponseWriter) {
-	Error(w, http.StatusBadRequest, "Unparsable JSON")
+func invalidJSON(w http.ResponseWriter, logger chronograf.Logger) {
+	Error(w, http.StatusBadRequest, "Unparsable JSON", logger)
 }
 
-func unknownErrorWithMessage(w http.ResponseWriter, err error) {
-	Error(w, http.StatusInternalServerError, fmt.Sprintf("Unknown error: %v", err))
+func unknownErrorWithMessage(w http.ResponseWriter, err error, logger chronograf.Logger) {
+	Error(w, http.StatusInternalServerError, fmt.Sprintf("Unknown error: %v", err), logger)
 }
 
-func notFound(w http.ResponseWriter, id int) {
-	Error(w, http.StatusNotFound, fmt.Sprintf("ID %d not found", id))
+func notFound(w http.ResponseWriter, id int, logger chronograf.Logger) {
+	Error(w, http.StatusNotFound, fmt.Sprintf("ID %d not found", id), logger)
 }
 
 func paramID(key string, r *http.Request) (int, error) {