+++ server/service.go
@@ -9,6 +9,14 @@ type Service struct {
 	ServersStore     chronograf.ServersStore
 	LayoutStore      chronograf.LayoutStore
 	AlertRulesStore  chronograf.AlertRulesStore
+	UsersStore       chronograf.UsersStore
 	TimeSeries       chronograf.TimeSeries
 	Logger           chronograf.Logger
+	UseAuth          bool
+}
+
+// ErrorMessage is the error response format for all service errors
+type ErrorMessage struct {
+	Code    int    `json:"code"`
+	Message string `json:"message"`
 }