+++ ui/src/auth/index.js
@@ -1,0 +1,2 @@
+import Login from './Login';
+export {Login};