+++ ui/src/side_nav/containers/SideNavApp.js
@@ -1,41 +1,38 @@
 import React, {PropTypes} from 'react';
+import {connect} from 'react-redux';
 import SideNav from '../components/SideNav';
 
-const {func, string} = PropTypes;
+const {func, string, shape} = PropTypes;
 const SideNavApp = React.createClass({
   propTypes: {
     currentLocation: string.isRequired,
     addFlashMessage: func.isRequired,
     sourceID: string.isRequired,
     explorationID: string,
-  },
-
-  contextTypes: {
-    canViewChronograf: PropTypes.bool,
-  },
-
-  getInitialState() {
-    return {
-      clusters: [],
-      clusterToUpdate: '',
-    };
+    me: shape({
+      email: string.isRequired,
+    }),
   },
 
   render() {
-    const {currentLocation, sourceID, explorationID} = this.props;
-    const {canViewChronograf} = this.context;
+    const {me, currentLocation, sourceID, explorationID} = this.props;
 
     return (
       <SideNav
         sourceID={sourceID}
-        isAdmin={true}
-        canViewChronograf={canViewChronograf}
         location={currentLocation}
         explorationID={explorationID}
+        me={me}
       />
     );
   },
 
 });
 
-export default SideNavApp;
+function mapStateToProps(state) {
+  return {
+    me: state.me,
+  };
+}
+
+export default connect(mapStateToProps)(SideNavApp);